Extra function for miniproject
2 Extra function has been written.
Showing the last 20 times when the toalett has been cleaned. The function works by pressing the 0x4 button the register the time. and switch on the switch 0x1 to show the list and scroll it down and up using buttons 0x1, 0x2. The function has written by Karoline Malko
Reconfigurating the texts [EMPTY], [BUSY] code was written by Garo Malko. The way it works by enabling the switch 0x8 by using the 1st, 2nd, 3th and 4th buttons on the chipkit.